updated hbmcmc to use externally supplied basis functions instead of on the fly quadtree
Former-commit-id: a2e371c25399e75b309ae01ececa3fb0cb38ac68

+++ acrg_hbmcmc/hbmcmc.py
@@ -252,8 +252,9 @@ def fixedbasisMCMC(species, sites, domain, meas_period, start_date,
                            start_date, end_date, outputname, outputpath,
                            basis_directory, country_file, fp_basis_case, country_unit_prefix)
     
-    # remove the temporary basis function directory
-    shutil.rmtree(tempdir)
+    if quadtree_basis is True:
+        # remove the temporary basis function directory
+        shutil.rmtree(tempdir)
     
     print("All done")
